+++ .private/scripts/resolvers/JsMinResolver.php
@@ -87,9 +87,9 @@ class JsMinResolver implements \framework\interfaces\IRequestResolver {
       $srcPath = "./$srcPath$_srcPath";
 
       // compile when: target file not exists, or source is newer
-      if ( !file_exists($dstPath) || @filemtime($srcPath) > @filemtime($dstPath) ) {
+      if ( file_exists($srcPath) && (!file_exists($dstPath) || @filemtime($srcPath) > @filemtime($dstPath)) ) {
         // empty results are ignored
-        $result = JSMin::minify(file_get_contents($srcPath));
+        $result = trim(@JSMin::minify(file_get_contents($srcPath)));
         if ( $result && !@file_put_contents($dstPath, $result) ) {
           Log::warn('Permission denied, unable to minify Javascript files.');
         }
